changed times interval & finer B-spline gird
the data cost is funny when starting the protocol from 1500 for a. It might be related to that linear ramp (though I don't see that on the model output). Instead we will run from t=3500 to all most the end of the protocol at t_end=14300.
Tuuned fine & medium Bs-pline grids to better approximate the transient of r. data cost for the entire time-series went down to 10^(-4), but the number of splines per segment increses to 13.